Control.Layers: stop relying on Browser `android` and `touch` properties
- `android` is useless here, as there are a lot of non-android touch devices (including desktop ones)
- `touch` is useless, as mouse can be used on touch-screen too

Most browsers produce compatibility mouseover for touch event, but not all,
so we listen for click also*.

But if both events occur together - then click can propagate to layers list and induce
unexpected input switching. To avoid this we defer `mouseover` action with setTimeout
(and cancel deferred action when it is not needed).
Note: more straight-forward way would be relying on e.sourceCapabilities.firesTouchEvents,
      but atm it has rather low support: https://caniuse.com/#feat=mdn-api_inputdevicecapabilities_firestouchevents

---
* small bonus: now one more usecase is covered
When Layers control created with {collapsed:false} option is collapsed explicitely
[using .collapse() method], it's now possible to expand it back with click.